+++ massiv-bench/bench/Stencil.hs
@@ -23,7 +23,7 @@ main = do
             [ env
                 (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
                 (bench "Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (average3x3Filter (Fill 0))))
+                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3Filter))
             , env
                 (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
                 (bench "Unsafe Array Ix2" .
@@ -31,18 +31,18 @@ main = do
             , env
                 (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
                 (bench "Convolve Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (average3x3FilterConv (Fill 0))))
+                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3FilterConv))
             ]
         , bgroup
             "Average Par"
             [ env
                 (return (computeAs U (arrDLightIx2 Par (tupleToIx2 t2))))
                 (bench "Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (average3x3Filter (Fill 0))))
+                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3Filter))
             , env
                 (return (computeAs U (arrDLightIx2 Par (tupleToIx2 t2))))
                 (bench "Convolve Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (average3x3FilterConv (Fill 0))))
+                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3FilterConv))
             ]
         ]
     -- , bgroup