+++ massiv/CHANGELOG.md
@@ -6,7 +6,8 @@
 * Breaking changes:
   * Create new `Data.Massiv.Array.Stencil.Unsafe` module and move `forStencilUnsafe` into it.
   * Rename of rank -> dimensions #25
-  * Removal `Eq` and `Ord` instances for `Value` #19
+    * Removal `Eq` and `Ord` instances for `Value` #19
+    * Move border resolution to `mapStencil` from `makeStencil`.
 
 # 0.1.6